% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the EvalTest Shiny application}
\usage{
run_app()
}
\value{
The function does not return a value; it launches a Shiny application.
}
\description{
This function starts the Shiny application included in the EvalTest package,
which aims to evaluate diagnostic tests performance.
}
\examples{
if (interactive()) {
  library(EvalTest)
  run_app()
}

}
